Move runtime helpers under Tapioca::Runtime
See https://github.com/Shopify/tapioca/pull/2387/files#r2398587417.
